fix subscription for async migration consumer
