+++ src/main/java/org/folio/services/kafka/KafkaProperties.java
@@ -3,7 +3,6 @@ package org.folio.services.kafka;
 import io.vertx.kafka.client.serialization.JsonObjectDeserializer;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.folio.kafka.SimpleConfigurationReader;
-import org.folio.services.kafka.topic.KafkaTopic;
 
 import java.util.HashMap;
 import java.util.List;
@@ -11,7 +10,6 @@ import java.util.Map;
 
 import static java.lang.Short.parseShort;
 import static java.lang.System.getenv;
-import static org.folio.Environment.environmentName;
 
 public final class KafkaProperties {
   private static String port = getenv().getOrDefault("KAFKA_PORT", "9092");
@@ -37,13 +35,14 @@ public final class KafkaProperties {
     return host;
   }
 
-  public static Map<String, String> getKafkaConsumerProperties(String tenantId, String groupId) {
+  public static Map<String, String> getKafkaConsumerProperties(String groupId) {
     Map<String, String> config = new HashMap<>();
     config.put("bootstrap.servers", KafkaProperties.getHost() + ":" + KafkaProperties.getPort());
     config.put("key.deserializer", StringDeserializer.class.getName());
     config.put("max.poll.records", SimpleConfigurationReader.getValue(List.of("kafka.consumer.max.poll.records", "spring.kafka.consumer.max-poll-records"), "100"));
     config.put("value.deserializer", JsonObjectDeserializer.class.getName());
     config.put("group.id", groupId);
+    config.put("metadata.max.age.ms", "15000");
     config.put("auto.offset.reset", "earliest");
     config.put("enable.auto.commit", "false");
     return config;
