+++ src/main/java/org/folio/services/kafka/topic/KafkaTopic.java
@@ -22,6 +22,10 @@ public class KafkaTopic {
     return forName("inventory.async-migration", tenantId, environmentName);
   }
 
+  public static KafkaTopic asyncMigration(String environmentName) {
+    return forName("inventory.async-migration", "\\w{1,}", environmentName);
+  }
+
   public static KafkaTopic boundWith(String tenantId, String environmentName) {
     return forName("inventory.bound-with", tenantId, environmentName);
   }
